refactor: combine Form into ApplyForm
