+++ neobank/src/App/pages/CreditCard/ApplyForm/ApplyForm.scss
@@ -103,3 +103,88 @@
   background-color: rgba(128, 128, 128, 0.2);
   width: 250px;
 }
+
+.form-h4 {
+  @include h4;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+
+  &-button {
+    align-self: flex-end;
+  }
+}
+
+.error-message {
+  color: red;
+}
+
+.form-grid {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-template-rows: repeat(2, auto);
+  gap: 10px;
+
+  @media (max-width: 1270px) {
+    grid-template-columns: repeat(3, 1fr);
+  }
+
+  @media (max-width: 968px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (max-width: 656px) {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
+
+.form-p {
+  @include p-primary;
+}
+
+.redstar {
+  @include p-primary;
+
+  line-height: 100%;
+  color: #f00;
+}
+
+.form-input {
+  border: 1px solid rgba(128, 128, 128, 0.2);
+  border-radius: 6px;
+  width: 284px;
+  height: 40px;
+  background-color: #f9f5e3;
+  background-repeat: no-repeat;
+  background-position: calc(100% - 17px) center;
+
+  @media (max-width: 352px) {
+    width: 230px;
+  }
+
+  &:focus {
+    border-color: rgb(104, 87, 197);
+    outline: none;
+  }
+
+  &::placeholder {
+    @include ubuntu-regular;
+    padding: 10px;
+    font-size: 14px;
+    line-height: 100%;
+    color: #808080;
+  }
+
+  &.success {
+    border: 1px solid green;
+    background-image: url("data:image/svg+xml,%3Csvg width='18' height='18' viewBox='0 0 18 18' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M9 18C13.9706 18 18 13.9706 18 9C18 4.02944 13.9706 0 9 0C4.02944 0 0 4.02944 0 9C0 13.9706 4.02944 18 9 18ZM8.76822 12.6402L13.7682 6.64018L12.2318 5.35982L7.9328 10.5186L5.70711 8.29289L4.29289 9.70711L7.29289 12.7071L8.0672 13.4814L8.76822 12.6402Z' fill='%23008000' fill-opacity='0.8'/%3E%3C/svg%3E");
+  }
+
+  &.error {
+    border: 1px solid red;
+    background-image: url("data:image/svg+xml,%3Csvg width='18' height='18' viewBox='0 0 18 18' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M18 9C18 13.9706 13.9706 18 9 18C4.02944 18 0 13.9706 0 9C0 4.02944 4.02944 0 9 0C13.9706 0 18 4.02944 18 9ZM4.29289 13.7071C3.90237 13.3166 3.90237 12.6834 4.29289 12.2929L7.58579 9L4.29289 5.70711C3.90237 5.31658 3.90237 4.68342 4.29289 4.29289C4.68342 3.90237 5.31658 3.90237 5.70711 4.29289L9 7.58579L12.2929 4.29289C12.6834 3.90237 13.3166 3.90237 13.7071 4.29289C14.0976 4.68342 14.0976 5.31658 13.7071 5.70711L10.4142 9L13.7071 12.2929C14.0976 12.6834 14.0976 13.3166 13.7071 13.7071C13.3166 14.0976 12.6834 14.0976 12.2929 13.7071L9 10.4142L5.70711 13.7071C5.31658 14.0976 4.68342 14.0976 4.29289 13.7071Z' fill='%23FF5631'/%3E%3C/svg%3E");
+  }
+}
