+++ neobank/src/App/pages/HomePage/Subscribe/Subscribe.tsx
@@ -13,29 +13,6 @@ const Subscribe = () => {
     }
   }, []);
 
-  // const handleSubmit = async (e) => {
-  //   e.preventDefault();
-  //   try {
-  //     if (process.env.NODE_ENV === 'development') {
-  //       await new Promise((resolve) =>
-  //         setTimeout(
-  //           () =>
-  //             resolve({
-  //               data: { message: 'Subscription successful' },
-  //             }),
-  //           500,
-  //         ),
-  //       );
-  //     } else {
-  //       await axios.post('/email', { email });
-  //     }
-  //     setSubscribed(true);
-  //     localStorage.setItem('isSubscribed', 'true');
-  //   } catch (error) {
-  //     console.error('Subscription failed', error);
-  //   }
-  // };
-
   interface SubscriptionResponse {
     data: {
       message: string;
@@ -47,12 +24,10 @@ const Subscribe = () => {
     try {
       let response: SubscriptionResponse;
       if (process.env.NODE_ENV === 'development') {
-        // Simulated response for development
         response = await new Promise((resolve) =>
           setTimeout(() => resolve({ data: { message: 'Subscription successful' } }), 500),
         );
       } else {
-        // Real API call for production
         response = await axios.post('http://localhost:8080/email', { email });
       }
